Detach Preferences into PreferencesMixin
